[FIX] hr_holidays: refuse button on time off request no rights
When you take a time off for which there is no need of validation, you can
refuse it, although you don't have any time off rights. This fixes the issue,
you have to be either a time off officer or the manager of the one who took the
time off.

closes odoo/odoo#170514

Task: 3981373
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>